requirements.txt
preprocessor==1.1.3
python-dotenv==0.19.2
tweepy==4.5.0
nltk~=3.6.7
spacy~=3.2.1